MQ plugin for Kafka: stats are all unsigned long

+++ trackrdrd/src/mq/kafka/mq_kafka.h
@@ -46,13 +46,13 @@ typedef struct kafka_wrk {
     int			err;
     char		reason[LINE_MAX]; /* errs from rdkafka callbacks    */
     char		errmsg[LINE_MAX]; /* thread-safe return from MQ_*() */
-    unsigned		seen;
-    unsigned		produced;
-    unsigned		delivered;
-    unsigned		failed;
-    unsigned		nokey;
-    unsigned		badkey;
-    unsigned		nodata;
+    unsigned long	seen;
+    unsigned long	produced;
+    unsigned long	delivered;
+    unsigned long	failed;
+    unsigned long	nokey;
+    unsigned long	badkey;
+    unsigned long	nodata;
 } kafka_wrk_t;
 
 kafka_wrk_t **workers;
